feat: cookies implement

+++ models/UserModel.php
@@ -2,12 +2,14 @@
 class UserModel extends Model {
 	private $_sessionName,
 			$_data,
-			$_isLoggedIn;
+			$_isLoggedIn,
+			$_cookieName;
 
     public function __construct($user = null) {
     	echo $this->_data;
         parent::__construct();
         $this->_sessionName = Config::get('session/session_name');
+        $this->_cookieName = Config::get('remember/cookie_name');
         if (!$user) {
         	if (Session::exists($this->_sessionName)) {
         		$user = Session::get($this->_sessionName);
@@ -61,23 +63,42 @@ class UserModel extends Model {
 		}
 	}
 
-	public function login($email, $password) {
-		$this->_data = $this->find($email);
-		$user = $this->_data;
-		if ($user) {
-			if ($user->password === Hash::make($password, $user->salt)) {
-				/* storing user id in session: $_SESSION[$session_name] = 3;*/
-				Session::put($this->_sessionName, $user->id);
-				return true;
-			} else {
-				echo '<br/> passwords dont match <br/>';
+	public function login($email = null, $password = null, $remember = false) {
+		if (!$email && !$password && $this->exists()) {
+			Session::put($this->_sessionName, $this->data()->id);
+		} else {
+			$user = $this->find($email);
+			if ($user) {
+				if ($user->password === Hash::make($password, $user->salt)) {
+					/* storing user id in session: $_SESSION[$session_name] = 3;*/
+					Session::put($this->_sessionName, $user->id);
+					if ($remember) {
+						/* storing in cookies */
+						$hashCheck = $this->_db->get('users_session', array('user_id', '=', $user->id));
+						if (!$hashCheck->count()) {
+							$hash = Hash::unique();
+							$this->_db->insert('users_session', array(
+								'user_id' => $user->id,
+								'hash' => $hash
+							));
+						} else {
+							$hash = $hashCheck->first()->hash;
+						}
+						Cookie::put($this->_cookieName, $hash, Config::get('remember/cookie_expiry'));
+					}
+					return true;
+				} else {
+					echo '<br/> passwords dont match <br/>';
+				}
 			}
 		}
 		return false;
 	}
 
 	public function logout() {
+		$this->_db->delete('users_session', array('user_id', '=', $this->data()->id));
 		Session::delete($this->_sessionName);
+		Cookie::delete($this->_cookieName);
 	}
 
 	public function data() {
@@ -86,6 +107,11 @@ class UserModel extends Model {
 
 	public function isLoggedIn() {
 		return $this->_isLoggedIn;
+	}
+
+	/*  */
+	public function exists() {
+		return (!empty($this->_data)) ? true : false;
 	}
 
 }
